PDF_Scraper/requirements.txt
pandas==2.0.3
flask==3.0.3
#python-doctr==0.9.0
numpy==1.24.3
img2table==1.3.0
openai==1.45.0
#langchain-community==0.3.0
langchain_openai==0.2.0
onnxtr[cpu]
#protobuf==3.20.3
#flatbuffers==24.3.25
onnxtr[cpu]
python-dotenv==1.0.1